Adds Skills
Skills are basically method modifiers. They jump on methods, and
and change their return values. For example, the Farsight skill
increases LOS by two, by intercepting the LOS distance method.

Skills can either used to change return values, or to trigger
callbacks. As proof of concept of that, the Vampirism skill restores
health when a unit calls #hit.

This commit also removes a lot of the basic configuration of classes
in favor of implementing it through skills.